+++ modules/localization/msf/local_tool/local_visualization/engine/visualization_engine.cc
@@ -235,7 +235,8 @@ void VisualizationEngine::Preprocess(const std::string &map_folder,
     return;
   }
 
-  GenerateMutiResolutionImages(map_bin_path, image_resolution_path.length(),
+  GenerateMutiResolutionImages(map_bin_path,
+                               static_cast<int>(image_resolution_path.length()),
                                image_visual_resolution_path_);
 }
 
@@ -266,8 +267,8 @@ void VisualizationEngine::Draw() {
 
   for (int i = 0; i < 3; ++i) {
     for (int j = 0; j < 3; ++j) {
-      subMat_[i][j].copyTo(
-          big_window_(cv::Rect(j * 1024, i * 1024, 1024, 1024)));
+      subMat_[i]
+             [j].copyTo(big_window_(cv::Rect(j * 1024, i * 1024, 1024, 1024)));
     }
   }
 
@@ -444,8 +445,8 @@ void VisualizationEngine::DrawStd(const cv::Point &bias) {
           unsigned char g = color_table[i % 3][1];
           unsigned char r = color_table[i % 3][2];
 
-          cv::Size size(std::sqrt(std[0]) * 200 + 1.0,
-                        std::sqrt(std[1]) * 200 + 1.0);
+          cv::Size size(static_cast<int>(std::sqrt(std[0]) * 200.0 + 1.0),
+                        static_cast<int>(std::sqrt(std[1]) * 200.0 + 1.0));
           cv::ellipse(big_window_, lt, size, 0, 0, 360, cv::Scalar(b, g, r), 2,
                       8);
         }
@@ -501,10 +502,9 @@ void VisualizationEngine::DrawLegend() {
     unsigned char b = color_table[i % 3][0];
     unsigned char g = color_table[i % 3][1];
     unsigned char r = color_table[i % 3][2];
-    cv::circle(
-        image_window_,
-        cv::Point(755, (15 + textSize.height) * (i + 1) - textSize.height / 2),
-        8, cv::Scalar(b, g, r), 3);
+    cv::circle(image_window_, cv::Point(755, (15 + textSize.height) * (i + 1) -
+                                                 textSize.height / 2),
+               8, cv::Scalar(b, g, r), 3);
   }
 }
 
@@ -596,8 +596,8 @@ void VisualizationEngine::GenerateMutiResolutionImages(
   int x_max = -1;
   int y_min = INT_MAX;
   int y_max = -1;
-  for (unsigned int i = 0; i < src_files.size(); ++i) {
-    int len = src_files[i].length();
+  for (size_t i = 0; i < src_files.size(); ++i) {
+    int len = static_cast<int>(src_files[i].length());
     // src_file example: image/000/north/50/00034661/00003386.png
     int y = std::stoi(src_files[i].substr(len - 21, 8));
     int x = std::stoi(src_files[i].substr(len - 12, 8));
@@ -739,10 +739,12 @@ void VisualizationEngine::CloudToMat(const Eigen::Affine3d &cur_pose,
   unsigned int img_width = map_param_.map_node_size_x;
   unsigned int img_height = map_param_.map_node_size_y;
   Eigen::Vector3d cen = car_pose_.translation();
-  cloud_img_lt_coord_[0] =
-      cen[0] - map_param_.map_resolutions[resolution_id_] * (img_width / 2.0f);
-  cloud_img_lt_coord_[1] =
-      cen[1] - map_param_.map_resolutions[resolution_id_] * (img_height / 2.0f);
+  cloud_img_lt_coord_[0] = cen[0] -
+                           map_param_.map_resolutions[resolution_id_] *
+                               (static_cast<float>(img_width) / 2.0f);
+  cloud_img_lt_coord_[1] = cen[1] -
+                           map_param_.map_resolutions[resolution_id_] *
+                               (static_cast<float>(img_height) / 2.0f);
 
   cloud_img_.setTo(cv::Scalar(0, 0, 0));
   cloud_img_mask_.setTo(cv::Scalar(0));
@@ -776,18 +778,18 @@ void VisualizationEngine::CoordToImageKey(const Eigen::Vector2d &coord,
   DCHECK_LT(resolution_id_, map_param_.map_resolutions.size());
   key->zone_id = zone_id_;
   int n = static_cast<int>((coord[0] - map_param_.map_min_x) /
-                           (map_param_.map_node_size_x *
+                           (static_cast<float>(map_param_.map_node_size_x) *
                             map_param_.map_resolutions[resolution_id_]));
   int m = static_cast<int>((coord[1] - map_param_.map_min_y) /
-                           (map_param_.map_node_size_y *
+                           (static_cast<float>(map_param_.map_node_size_y) *
                             map_param_.map_resolutions[resolution_id_]));
   int max_n = static_cast<int>((map_param_.map_max_x - map_param_.map_min_x) /
-                               (map_param_.map_node_size_x *
+                               (static_cast<float>(map_param_.map_node_size_x) *
                                 map_param_.map_resolutions[resolution_id_]));
-  int max_m =
-      static_cast<unsigned int>((map_param_.map_max_y - map_param_.map_min_y) /
-                                (map_param_.map_node_size_y *
-                                 map_param_.map_resolutions[resolution_id_]));
+  int max_m = static_cast<unsigned int>(
+      (map_param_.map_max_y - map_param_.map_min_y) /
+      (static_cast<float>(map_param_.map_node_size_y) *
+       map_param_.map_resolutions[resolution_id_]));
 
   if (n >= 0 && m >= 0 && n < max_n && m < max_m) {
     key->node_north_id = m;
@@ -862,7 +864,8 @@ void VisualizationEngine::RotateImg(const cv::Mat &in_img, cv::Mat *out_img,
   in_img.copyTo(
       mat_tem(cv::Rect(width / 2 - in_img.cols / 2, width / 2 - in_img.rows / 2,
                        in_img.cols, in_img.rows)));
-  cv::Point2f pt(width / 2.0f, width / 2.0f);
+  cv::Point2f pt(static_cast<float>(width) / 2.0f,
+                 static_cast<float>(width) / 2.0f);
   cv::Mat rotation_mat = cv::getRotationMatrix2D(pt, angle, 1.0);
   cv::warpAffine(mat_tem, *out_img, rotation_mat, cv::Size(width, width));
 }
